alfred: fix error code from unix_sock_req_data_reply
The error code of unix_sock_req_data_reply() was stored in an uint16_t
before being returned from the int-returning function. A failed write to
the client socket therefore returned 65535 instead of -1.

Fixes: 4a77be458853 ("alfred: handle failing write() in unix sockets")

+++ unix_sock.c
@@ -230,7 +230,7 @@ static int unix_sock_req_data_reply(struct globals *globals, int client_sock,
 	struct hash_it_t *hashit = NULL;
 	uint8_t buf[MAX_PAYLOAD];
 	uint16_t seqno = 0;
-	uint16_t ret = 0;
+	int ret = 0;
 	int len;
 
 	/* send some data back through the unix socket */
